Renamed the tables directory to learning.
- As suggested by Oliver, I have renamed the tables directory to
  learning, which is probably a better name for the current files in
  the directory.
