+++ code/factorbase/src/main/java/ca/sfu/cs/factorbase/learning/CPGenerator.java
@@ -1,4 +1,4 @@
-package ca.sfu.cs.factorbase.tables;
+package ca.sfu.cs.factorbase.learning;
 
 import ca.sfu.cs.common.Configuration.Config;
